refactor: split classifier exec & training
